Move BeatmapManagerWorkingBeatmap to its own file
